Pinterest/requirements.txt
urllib3
py3-pinterest